Green test - Statement object returns a single transaction
Created a new Statement class responsible for presenting the statement in order to reduce the responsibilty on the Account class.

Have not changed any tests or functionality of the Account class yet, but by using doubles and stubs when testing the Statement class, I am able to test the Statement class in isolation. And this will also allow me to later implement dependency injection